chore(admin): convert ProfilePage and re-use server driven types (#18593)
* chore(admin): convert ProfilePage and re-use server driven types

* chore: fix tests

* chore: fix double export from types

* fix: type isSSOLocked endpoint
