+++ packages/core/admin/admin/src/components/tests/RBACProvider.test.tsx
@@ -111,6 +111,7 @@ describe('RBACReducer', () => {
       };
 
       expect(
+        // @ts-expect-error – admin-test-utils permission types are incorrect
         RBACReducer(state, setPermissionsAction(fixtures.permissions.contentManager))
           .collectionTypesRelatedPermissions
       ).toEqual(expected);
