+++ packages/core/admin/admin/src/pages/SettingsPage/pages/Roles/hooks/tests/useAdminRolePermissions.test.ts
@@ -26,6 +26,7 @@ describe('useAdminRolePermissions', () => {
   });
 
   test('forwards all query params except `id`', async () => {
+    // @ts-expect-error - test purposes
     const { result } = setup({ id: 1, filters: 'param' });
 
     await waitFor(() => expect(result.current.isLoading).toBe(false));
