+++ packages/core/admin/shared/contracts/admin.ts
@@ -1,0 +1,24 @@
+import { errors } from '@strapi/utils';
+
+/**
+ * /information - get project information
+ */
+export declare namespace Information {
+  export interface Request {
+    body: {};
+    query: {};
+  }
+  export interface Response {
+    data: {
+      currentEnvironment: string;
+      autoReload: boolean;
+      strapiVersion: string | null;
+      dependencies: Record<string, string>;
+      projectId: string | null;
+      nodeVersion: string;
+      communityEdition: boolean;
+      useYarn: boolean;
+    };
+    error?: errors.ApplicationError;
+  }
+}
