+++ packages/core/admin/shared/contracts/providers.ts
@@ -1,0 +1,17 @@
+import { errors } from '@strapi/utils';
+
+/**
+ * /providers/isSSOLocked
+ */
+export declare namespace IsSSOLocked {
+  export interface Request {
+    body: {};
+    query: {};
+  }
+  export interface Response {
+    data: {
+      isSSOLocked: boolean;
+    };
+    error?: errors.ApplicationError;
+  }
+}
