+++ packages/core/helper-plugin/src/utils/tests/hasPermissions.test.ts
@@ -16,6 +16,8 @@ const hasPermissionsTestData: Record<string, Record<string, Permission[]>> = {
         subject: null,
         properties: {},
         conditions: [],
+        actionParameters: {},
+        id: 1,
       },
 
       // Admin webhooks
@@ -24,24 +26,32 @@ const hasPermissionsTestData: Record<string, Record<string, Permission[]>> = {
         subject: null,
         properties: {},
         conditions: [],
+        actionParameters: {},
+        id: 2,
       },
       {
         action: 'admin::webhooks.read',
         subject: null,
         properties: {},
         conditions: [],
+        actionParameters: {},
+        id: 3,
       },
       {
         action: 'admin::webhooks.update',
         subject: null,
         properties: {},
         conditions: [],
+        actionParameters: {},
+        id: 4,
       },
       {
         action: 'admin::webhooks.delete',
         subject: null,
         properties: {},
         conditions: [],
+        actionParameters: {},
+        id: 5,
       },
 
       // Admin users
@@ -50,24 +60,32 @@ const hasPermissionsTestData: Record<string, Record<string, Permission[]>> = {
         subject: null,
         properties: {},
         conditions: [],
+        actionParameters: {},
+        id: 6,
       },
       {
         action: 'admin::users.read',
         subject: null,
         properties: {},
         conditions: [],
+        actionParameters: {},
+        id: 7,
       },
       {
         action: 'admin::users.update',
         subject: null,
         properties: {},
         conditions: [],
+        actionParameters: {},
+        id: 8,
       },
       {
         action: 'admin::users.delete',
         subject: null,
         properties: {},
         conditions: [],
+        actionParameters: {},
+        id: 9,
       },
 
       // Admin roles
@@ -76,24 +94,32 @@ const hasPermissionsTestData: Record<string, Record<string, Permission[]>> = {
         subject: null,
         properties: {},
         conditions: [],
+        actionParameters: {},
+        id: 10,
       },
       {
         action: 'admin::roles.read',
         subject: null,
         properties: {},
         conditions: [],
+        actionParameters: {},
+        id: 11,
       },
       {
         action: 'admin::roles.update',
         subject: null,
         properties: {},
         conditions: [],
+        actionParameters: {},
+        id: 12,
       },
       {
         action: 'admin::roles.delete',
         subject: null,
         properties: {},
         conditions: [],
+        actionParameters: {},
+        id: 13,
       },
 
       // Content type builder
@@ -101,6 +127,9 @@ const hasPermissionsTestData: Record<string, Record<string, Permission[]>> = {
         action: 'plugin::content-type-builder.read',
         subject: null,
         properties: {},
+        conditions: [],
+        actionParameters: {},
+        id: 14,
       },
 
       // Documentation plugin
@@ -108,16 +137,25 @@ const hasPermissionsTestData: Record<string, Record<string, Permission[]>> = {
         action: 'plugin::documentation.read',
         subject: null,
         properties: {},
+        conditions: [],
+        actionParameters: {},
+        id: 15,
       },
       {
         action: 'plugin::documentation.settings.update',
         subject: null,
         properties: {},
+        conditions: [],
+        actionParameters: {},
+        id: 16,
       },
       {
         action: 'plugin::documentation.settings.regenerate',
         subject: null,
         properties: {},
+        conditions: [],
+        actionParameters: {},
+        id: 17,
       },
 
       // Upload plugin
@@ -125,26 +163,41 @@ const hasPermissionsTestData: Record<string, Record<string, Permission[]>> = {
         action: 'plugin::upload.read',
         subject: null,
         properties: {},
+        conditions: [],
+        actionParameters: {},
+        id: 18,
       },
       {
         action: 'plugin::upload.assets.create',
         subject: null,
         properties: {},
+        conditions: [],
+        actionParameters: {},
+        id: 19,
       },
       {
         action: 'plugin::upload.assets.update',
         subject: null,
         properties: {},
+        conditions: [],
+        actionParameters: {},
+        id: 20,
       },
       {
         action: 'plugin::upload.assets.dowload',
         subject: null,
         properties: {},
+        conditions: [],
+        actionParameters: {},
+        id: 21,
       },
       {
         action: 'plugin::upload.assets.copy-link',
         subject: null,
         properties: {},
+        conditions: [],
+        actionParameters: {},
+        id: 22,
       },
 
       // Users-permissions
@@ -152,51 +205,81 @@ const hasPermissionsTestData: Record<string, Record<string, Permission[]>> = {
         action: 'plugin::users-permissions.roles.create',
         subject: null,
         properties: {},
+        conditions: [],
+        actionParameters: {},
+        id: 23,
       },
       {
         action: 'plugin::users-permissions.roles.read',
         subject: null,
         properties: {},
+        conditions: [],
+        actionParameters: {},
+        id: 24,
       },
       {
         action: 'plugin::users-permissions.roles.update',
         subject: null,
         properties: {},
+        conditions: [],
+        actionParameters: {},
+        id: 25,
       },
       {
         action: 'plugin::users-permissions.roles.delete',
         subject: null,
         properties: {},
+        conditions: [],
+        actionParameters: {},
+        id: 26,
       },
       {
         action: 'plugin::users-permissions.email-templates.read',
         subject: null,
         properties: {},
+        conditions: [],
+        actionParameters: {},
+        id: 27,
       },
       {
         action: 'plugin::users-permissions.email-templates.update',
         subject: null,
         properties: {},
+        conditions: [],
+        actionParameters: {},
+        id: 28,
       },
       {
         action: 'plugin::users-permissions.providers.read',
         subject: null,
         properties: {},
+        conditions: [],
+        actionParameters: {},
+        id: 29,
       },
       {
         action: 'plugin::users-permissions.providers.update',
         subject: null,
         properties: {},
+        conditions: [],
+        actionParameters: {},
+        id: 30,
       },
       {
         action: 'plugin::users-permissions.advanced-settings.read',
         subject: null,
         properties: {},
+        conditions: [],
+        actionParameters: {},
+        id: 31,
       },
       {
         action: 'plugin::users-permissions.advanced-settings.update',
         subject: null,
         properties: {},
+        conditions: [],
+        actionParameters: {},
+        id: 32,
       },
     ],
     user2: [
@@ -206,24 +289,32 @@ const hasPermissionsTestData: Record<string, Record<string, Permission[]>> = {
         subject: null,
         properties: {},
         conditions: [],
+        actionParameters: {},
+        id: 33,
       },
       {
         action: 'admin::webhooks.read',
         subject: null,
         properties: {},
         conditions: [],
+        actionParameters: {},
+        id: 34,
       },
       {
         action: 'admin::webhooks.update',
         subject: null,
         properties: {},
         conditions: [],
+        actionParameters: {},
+        id: 35,
       },
       {
         action: 'admin::webhooks.delete',
         subject: null,
         properties: {},
         conditions: [],
+        actionParameters: {},
+        id: 36,
       },
 
       // Admin users
@@ -232,24 +323,32 @@ const hasPermissionsTestData: Record<string, Record<string, Permission[]>> = {
         subject: null,
         properties: {},
         conditions: [],
+        actionParameters: {},
+        id: 37,
       },
       {
         action: 'admin::users.read',
         subject: null,
         properties: {},
         conditions: [],
+        actionParameters: {},
+        id: 38,
       },
       {
         action: 'admin::users.update',
         subject: null,
         properties: {},
         conditions: [],
+        actionParameters: {},
+        id: 39,
       },
       {
         action: 'admin::users.delete',
         subject: null,
         properties: {},
         conditions: [],
+        actionParameters: {},
+        id: 40,
       },
 
       // Admin roles
@@ -258,24 +357,32 @@ const hasPermissionsTestData: Record<string, Record<string, Permission[]>> = {
         subject: null,
         properties: {},
         conditions: [],
+        actionParameters: {},
+        id: 41,
       },
       {
         action: 'admin::roles.read',
         subject: null,
         properties: {},
         conditions: [],
+        actionParameters: {},
+        id: 42,
       },
       {
         action: 'admin::roles.update',
         subject: null,
         properties: {},
         conditions: [],
+        actionParameters: {},
+        id: 43,
       },
       {
         action: 'admin::roles.delete',
         subject: null,
         properties: {},
         conditions: [],
+        actionParameters: {},
+        id: 44,
       },
 
       // Content type builder
@@ -284,6 +391,8 @@ const hasPermissionsTestData: Record<string, Record<string, Permission[]>> = {
         subject: null,
         properties: {},
         conditions: [],
+        actionParameters: {},
+        id: 45,
       },
 
       // Upload plugin
@@ -292,30 +401,40 @@ const hasPermissionsTestData: Record<string, Record<string, Permission[]>> = {
         subject: null,
         properties: {},
         conditions: [],
+        actionParameters: {},
+        id: 46,
       },
       {
         action: 'plugin::upload.assets.create',
         subject: null,
         properties: {},
         conditions: [],
+        actionParameters: {},
+        id: 47,
       },
       {
         action: 'plugin::upload.assets.update',
         subject: null,
         properties: {},
         conditions: [],
+        actionParameters: {},
+        id: 48,
       },
       {
         action: 'plugin::upload.assets.dowload',
         subject: null,
         properties: {},
         conditions: [],
+        actionParameters: {},
+        id: 49,
       },
       {
         action: 'plugin::upload.assets.copy-link',
         subject: null,
         properties: {},
         conditions: [],
+        actionParameters: {},
+        id: 50,
       },
 
       // Users-permissions
@@ -324,86 +443,229 @@ const hasPermissionsTestData: Record<string, Record<string, Permission[]>> = {
         subject: null,
         properties: {},
         conditions: [],
+        actionParameters: {},
+        id: 51,
       },
       {
         action: 'plugin::users-permissions.roles.read',
         subject: null,
         properties: {},
         conditions: [],
+        actionParameters: {},
+        id: 51,
       },
       {
         action: 'plugin::users-permissions.roles.update',
         subject: null,
         properties: {},
         conditions: [],
+        actionParameters: {},
+        id: 52,
       },
       {
         action: 'plugin::users-permissions.roles.delete',
         subject: null,
         properties: {},
         conditions: [],
+        actionParameters: {},
+        id: 53,
       },
       {
         action: 'plugin::users-permissions.email-templates.read',
         subject: null,
         properties: {},
         conditions: [],
+        actionParameters: {},
+        id: 54,
       },
       {
         action: 'plugin::users-permissions.email-templates.update',
         subject: null,
         properties: {},
         conditions: [],
+        actionParameters: {},
+        id: 55,
       },
       {
         action: 'plugin::users-permissions.providers.read',
         subject: null,
         properties: {},
         conditions: [],
+        actionParameters: {},
+        id: 56,
       },
       {
         action: 'plugin::users-permissions.providers.update',
         subject: null,
         properties: {},
         conditions: [],
+        actionParameters: {},
+        id: 57,
       },
       {
         action: 'plugin::users-permissions.advanced-settings.read',
         subject: null,
         properties: {},
         conditions: [],
+        actionParameters: {},
+        id: 58,
       },
       {
         action: 'plugin::users-permissions.advanced-settings.update',
         subject: null,
         properties: {},
         conditions: [],
+        actionParameters: {},
+        id: 59,
       },
     ],
   },
   permissionsToCheck: {
-    listPlugins: [{ action: 'admin::marketplace.read', subject: null }],
-    marketplace: [{ action: 'admin::marketplace.read', subject: null }],
+    listPlugins: [
+      {
+        action: 'admin::marketplace.read',
+        subject: null,
+        properties: {},
+        actionParameters: {},
+        conditions: [],
+        id: 1,
+      },
+    ],
+    marketplace: [
+      {
+        action: 'admin::marketplace.read',
+        subject: null,
+        properties: {},
+        actionParameters: {},
+        conditions: [],
+        id: 2,
+      },
+    ],
     settings: [
       // webhooks
-      { action: 'admin::webhook.create', subject: null },
-      { action: 'admin::webhook.read', subject: null },
-      { action: 'admin::webhook.update', subject: null },
-      { action: 'admin::webhook.delete', subject: null },
+      {
+        action: 'admin::webhook.create',
+        subject: null,
+        properties: {},
+        actionParameters: {},
+        conditions: [],
+        id: 3,
+      },
+      {
+        action: 'admin::webhook.read',
+        subject: null,
+        properties: {},
+        actionParameters: {},
+        conditions: [],
+        id: 4,
+      },
+      {
+        action: 'admin::webhook.update',
+        subject: null,
+        properties: {},
+        actionParameters: {},
+        conditions: [],
+        id: 5,
+      },
+      {
+        action: 'admin::webhook.delete',
+        subject: null,
+        properties: {},
+        actionParameters: {},
+        conditions: [],
+        id: 6,
+      },
       // users
-      { action: 'admin::users.create', subject: null },
-      { action: 'admin::users.read', subject: null },
-      { action: 'admin::users.update', subject: null },
-      { action: 'admin::users.delete', subject: null },
+      {
+        action: 'admin::users.create',
+        subject: null,
+        properties: {},
+        actionParameters: {},
+        conditions: [],
+        id: 7,
+      },
+      {
+        action: 'admin::users.read',
+        subject: null,
+        properties: {},
+        actionParameters: {},
+        conditions: [],
+        id: 8,
+      },
+      {
+        action: 'admin::users.update',
+        subject: null,
+        properties: {},
+        actionParameters: {},
+        conditions: [],
+        id: 9,
+      },
+      {
+        action: 'admin::users.delete',
+        subject: null,
+        properties: {},
+        actionParameters: {},
+        conditions: [],
+        id: 10,
+      },
       // roles
-      { action: 'admin::roles.create', subject: null },
-      { action: 'admin::roles.update', subject: null },
-      { action: 'admin::roles.read', subject: null },
-      { action: 'admin::roles.delete', subject: null },
+      {
+        action: 'admin::roles.create',
+        subject: null,
+        properties: {},
+        actionParameters: {},
+        conditions: [],
+        id: 11,
+      },
+      {
+        action: 'admin::roles.update',
+        subject: null,
+        properties: {},
+        actionParameters: {},
+        conditions: [],
+        id: 12,
+      },
+      {
+        action: 'admin::roles.read',
+        subject: null,
+        properties: {},
+        actionParameters: {},
+        conditions: [],
+        id: 13,
+      },
+      {
+        action: 'admin::roles.delete',
+        subject: null,
+        properties: {},
+        actionParameters: {},
+        conditions: [],
+        id: 14,
+      },
       // media library
-      { action: 'plugin::upload.read', subject: null },
-      { action: 'plugin::upload.assets.create', subject: null },
-      { action: 'plugin::upload.assets.update', subject: null },
+      {
+        action: 'plugin::upload.read',
+        subject: null,
+        properties: {},
+        actionParameters: {},
+        conditions: [],
+        id: 15,
+      },
+      {
+        action: 'plugin::upload.assets.create',
+        subject: null,
+        properties: {},
+        actionParameters: {},
+        conditions: [],
+        id: 16,
+      },
+      {
+        action: 'plugin::upload.assets.update',
+        subject: null,
+        properties: {},
+        actionParameters: {},
+        conditions: [],
+        id: 17,
+      },
     ],
   },
 };
@@ -433,6 +695,8 @@ describe('hasPermissions', () => {
           subject: null,
           properties: {},
           conditions: [],
+          id: 1,
+          actionParameters: {},
         },
       ]);
     });
@@ -448,18 +712,24 @@ describe('hasPermissions', () => {
             fields: ['test'],
           },
           conditions: [],
+          actionParameters: {},
+          id: 1,
         },
         {
           action: 'admin::webhooks.delete',
           subject: null,
           properties: {},
           conditions: ['customCondition'],
+          actionParameters: {},
+          id: 2,
         },
         {
           action: 'a',
           subject: 'b',
           properties: {},
           conditions: ['customCondition'],
+          actionParameters: {},
+          id: 3,
         },
       ];
 
@@ -498,10 +768,14 @@ describe('hasPermissions', () => {
 
     it('should return false no permission is matching', async () => {
       const data = hasPermissionsTestData.userPermissions.user1;
-      const dataToCheck = [
+      const dataToCheck: Permission[] = [
         {
+          id: 1,
           action: 'something',
           subject: 'something',
+          properties: {},
+          actionParameters: {},
+          conditions: [],
         },
       ];
 
@@ -517,12 +791,14 @@ describe('hasPermissions', () => {
     });
 
     it('should return false if there is no condition in the array of permissions', () => {
-      const data = [
+      const data: Permission[] = [
         {
+          id: 1,
           action: 'admin::marketplace.read',
           subject: null,
           properties: {},
           conditions: [],
+          actionParameters: {},
         },
       ];
 
@@ -530,23 +806,30 @@ describe('hasPermissions', () => {
     });
 
     it('should return false if there is at least one item that does not have a condition in the array of permissions', () => {
-      const data = [
+      const data: Permission[] = [
         {
+          id: 1,
           action: 'admin::marketplace.read',
           subject: null,
           properties: {},
           conditions: [],
+          actionParameters: {},
         },
         {
+          id: 2,
           action: 'admin::webhooks.delete',
           subject: null,
           properties: {},
           conditions: ['customCondition'],
+          actionParameters: {},
         },
         {
+          id: 3,
           action: 'admin::webhooks.create',
           subject: null,
           properties: {},
+          actionParameters: {},
+          conditions: [],
         },
       ];
 
@@ -554,24 +837,30 @@ describe('hasPermissions', () => {
     });
 
     it('should return true otherwise', () => {
-      const data = [
+      const data: Permission[] = [
         {
+          id: 1,
           action: 'admin::marketplace.read',
           subject: null,
           properties: {},
           conditions: ['test'],
+          actionParameters: {},
         },
         {
+          id: 2,
           action: 'admin::webhooks.delete',
           subject: null,
           properties: {},
           conditions: ['customCondition'],
+          actionParameters: {},
         },
         {
+          id: 3,
           action: 'admin::webhooks.create',
           subject: null,
           properties: {},
           conditions: ['test'],
+          actionParameters: {},
         },
       ];
